No longer create links for non-existent config files

+++ install.sh
@@ -19,8 +19,10 @@ read -n 1 bash_config                                              # Read input 
 if [[ $bash_config == "Y" ]] || [[ $bash_config == "y" ]]; then     # Allow user to choose which configurations they would like
     for i in .bashrc .bash_profile .bash_completion                 # More options may be added here in the future
     do
-        mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null          # Create a backup of the exisiting config file
-        ln -s $PWD/$i $INSTALLDIR/$i;                                           # Create a symlink to the new file
+	if [[ -f $INSTALLDIR/$i ]] || [[ -d $INSTALLDIR/$i ]]; then
+		mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null          # Create a backup of the exisiting config file
+		ln -s $PWD/$i $INSTALLDIR/$i;                                           # Create a symlink to the new file
+	fi
     done
 fi
 
@@ -30,10 +32,12 @@ echo -n "Set up Z-Shell configuration files? (y/n)"
 read -n 1 zsh_config
 
 if [[ $zsh_config == "Y" ]] || [[ $zsh_config == "y" ]]; then 
-    for i in .zshrc .z_profile .zsh_completion .oh-my-zsh
+    for i in .zshrc .z_profile .zsh_completion #.oh-my-zsh  # Problem with .oh-my-zsh
     do
-        mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null
-        ln -s $PWD/$i $INSTALLDIR/$i;
+	if [[ -f $INSTALLDIR/$i ]] || [[ -d $INSTALLDIR/$i ]]; then
+		mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null
+		ln -s $PWD/$i $INSTALLDIR/$i;
+	fi
     done
 fi
 
@@ -43,14 +47,16 @@ echo -n "Set up Vim configuration files? (y/n)"
 read -n 1 vim_config
 
 if [[ $vim_config == "Y" ]] || [[ $vim_config == "y" ]]; then 
-    for i in .vimrc .vim _vimrc
+    for i in .vimrc .vim .vrapperrc
     do
-        mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null
-        ln -s $PWD/$i $INSTALLDIR/$i;
+	if [[ -f $INSTALLDIR/$i ]] || [[ -d $INSTALLDIR/$i ]]; then
+		mv $INSTALLDIR/$i $INSTALLDIR/$i.backup 2>/dev/null
+		ln -s $PWD/$i $INSTALLDIR/$i;
+	fi
     done
 fi
 
 echo;
 echo "Process Complete"
-echo "Any existing files were renamed to <file>.backup"
+echo "Any previously existing files were renamed to <file>.backup"
 exit 0
